requirements.txt
aiocache==0.12.2
aiofiles==0.8.0
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.17.0
asyncpraw==7.7.1
asyncprawcore==2.4.0
attrs==23.2.0
black==24.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
frozenlist==1.4.1
idna==3.6
isort==5.13.2
multidict==6.0.5
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
python-dotenv==1.0.1
requests==2.31.0
typing_extensions==4.9.0
update-checker==0.18.0
urllib3==2.2.1
yarl==1.9.4
